turbine/validation/requirements.txt
# PTRM Validation Suite Dependencies
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
scikit-image>=0.19.0
Pillow>=9.0.0
tifffile>=2022.0.0
